+++ src/main/c/cext/cext_constants.c
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2019 Oracle and/or its affiliates. All rights reserved. This
+ * Copyright (c) 2020 Oracle and/or its affiliates. All rights reserved. This
  * code is released under a tri EPL/GPL/LGPL license. You can use it,
  * redistribute it and/or modify it under the terms of the:
  *
@@ -104,6 +104,10 @@ VALUE rb_tr_get_Numeric(void) {
 
 VALUE rb_tr_get_Object(void) {
   return RUBY_CEXT_INVOKE("rb_cObject");
+}
+
+VALUE rb_tr_get_Process(void) {
+  return RUBY_CEXT_INVOKE("rb_mProcess");
 }
 
 VALUE rb_tr_get_Range(void) {