+++ CHANGELOG.md
@@ -74,6 +74,7 @@ Compatibility:
 * `RSTRING_END()` now always returns a native pointer.
 * Removed `register` specifier for `rb_mem_clear()` (#1924).
 * Implemented `Thread::Backtrace::Locations#base_label` (#1920).
+* Implemented `rb_mProcess` (#1936).
 
 Changes: